requirements.txt
fastapi==0.63.0
uvicorn==0.11.8
motor==2.2.0
pydantic==1.7.3
PyJWT==1.7.1 
python-decouple==3.3
pymongo~=3.11.3
torch~=1.8.0
numpy~=1.20.1
Flask~=1.1.2
scipy~=1.6.1
torchaudio
google-api-python-client
progressbar
requests
pysrt
python-multipart
